get rid of unnecessary Provider concept
