+++ src/test/java/org/javarosa/test_utils/MockTimezoneProvider.java
@@ -1,0 +1,21 @@
+package org.javarosa.test_utils;
+
+import org.javarosa.core.model.utils.TimezoneProvider;
+
+/**
+ * Created by amstone326 on 1/8/18.
+ */
+
+public class MockTimezoneProvider extends TimezoneProvider {
+
+    private int offsetMillis;
+
+    public void setOffset(int offset) {
+        this.offsetMillis = offset;
+    }
+
+    @Override
+    public int getTimezoneOffsetMillis() {
+        return offsetMillis;
+    }
+}
